requirements.txt
langchain==0.3.0
faiss-cpu==1.8.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-huggingface==0.1.0
langchain-groq== 0.2.0
PyMuPDF==1.24.5
tiktoken==0.7.0
bs4==0.0.2
fastembed==0.4.2
tavily-python==0.5.0